apps/ml/pyproject.toml
[project]
name = "desain-ml"
version = "0.0.1"
description = "Forecasting & anomaly detection sidecar for DESAIN POS"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pyjwt>=2.9.0",
    "numpy>=2.1.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.5.0",
    "prophet>=1.1.5",
    "statsmodels>=0.14.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "ANN", "RUF"]
ignore = ["ANN101", "ANN102"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
